Add solution for Search Sequence in 2d array problem (based on the backtracking class):
Given A is a matrix and S is an array of integers, we say S occurs in A if you can start from
some entry in A and traverse adjacent entries in A in the order prescribed by S.
Adjacent entries are top, bottom, left and right ones.
It is possible to visit an entry in the matrix A more than once.

 For example, if A =
 [1 2 3
 3 4 5
 5 6 7]
 S = [1 3 4 6]
 Then S is in A, because A[0][0] = 1, A[1][0] = 3, A[1][1] = 4, A[2][1] = 6